requirements.txt
pyautogui
keyboard
opencv-python
pytesseract
Pillow
tkinter